make FullRay inherit from ray
