+++ src/main/maths/FullRay.java
@@ -1,34 +1,18 @@
 package main.maths;
 
-public class FullRay {
+public class FullRay extends Ray {
 
     private Vector3 origin;
 
     private Vector3 direction;
 
     public FullRay(Vector3 direction, Vector3 origin) {
-        this.direction = direction.normalise();
-        this.origin = origin;
+        super(direction, origin);
     }
 
     //some overloaded constructors to reduce boilerplate code
     public FullRay(double dirX, double dirY, double dirZ, double originX, double originY, double originZ) {
-        this.direction = new Vector3(dirX, dirY, dirZ).normalise();
-        this.origin = new Vector3(originX, originY, originZ);
-    }
-
-
-    public Vector3 getOrigin() {
-        return origin;
-    }
-
-    public Vector3 getDirection() {
-        return direction;
-    }
-
-    //returns a point that's the specified distance away from the origin and lies on the ray.g
-    public Vector3 getPointAlongRay(double distance) {
-        return origin.add(direction.multi(distance));
+        super(dirX, dirY, dirZ, originX, originY, originZ);
     }
 
 }
